refactor: - optimize apollo-provider - add footer
