+++ src/app/auth/layout.tsx
@@ -1,4 +1,5 @@
 import Logo from "@/components/logo";
+import ApolloProvider from "@/providers/apollo-provider";
 import Link from "next/link";
 
 interface AuthLayoutProps {
@@ -9,10 +10,12 @@ export default function AuthLayout({ children }: AuthLayoutProps) {
   return (
     <div className="min-h-screen flex items-center justify-center bg-muted/30 p-4">
       <div className="w-full max-w-md space-y-4 bg-background p-8 rounded-lg shadow-sm border">
-        <Link href="/">
-          <Logo />
-        </Link>
-        {children}
+        <ApolloProvider>
+          <Link href="/">
+            <Logo />
+          </Link>
+          {children}
+        </ApolloProvider>
       </div>
     </div>
   );
