+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import "./globals.css";
+import Footer from "@/components/footer";
 import { Toaster } from "@/components/ui/toaster";
-
 export const metadata: Metadata = {
   title: "YOUR MUSIC",
   description: "Convert your playlist to other platforms",
@@ -16,6 +16,7 @@ export default async function RootLayout({
     <html lang="en">
       <body>
         {children}
+        <Footer />
         <Toaster />
       </body>
     </html>
