+++ src/app/mypage/layout.tsx
@@ -1,4 +1,5 @@
 import Navbar from "@/components/nav-bar";
+import ApolloProvider from "@/providers/apollo-provider";
 
 export default function MyPageLayout({
   children,
@@ -6,9 +7,11 @@ export default function MyPageLayout({
   children: React.ReactNode;
 }) {
   return (
-    <div className="min-h-screen frutiger-aero-bg">
-      <Navbar />
-      <div className="container mx-auto px-4 py-8">{children}</div>
+    <div className="min-h-screen">
+      <ApolloProvider>
+        <Navbar />
+        <div className="container mx-auto px-4 py-8">{children}</div>
+      </ApolloProvider>
     </div>
   );
 }
