+++ src/app/playlists/[playlist-id]/_components/playlist.tsx
@@ -2,16 +2,14 @@
 
 import { Download } from "lucide-react";
 import { Button } from "@/components/ui/button";
-import { SongTable } from "@/app/playlists/_components/song-table";
 import {
   ConvertToSpotifyPlaylistButton,
   ConvertToYoutubePlaylistButton,
 } from "@/app/playlists/_components/buttons";
 import { useToast } from "@/hooks/use-toast";
 import { useAuthStore } from "@/store/auth-store";
-import { useSanitizedData } from "@/hooks/use-sanitizedata";
 import { Title } from "./title";
-import { GetPlaylistQuery } from "@/graphql/operations";
+import { PlaylistJson } from "@/graphql/types";
 
 const HEADERS = ["Title", "Artist", "Album"];
 const BOM = "\uFEFF";
@@ -22,28 +20,27 @@ const sanitizeFileName = (name: string): string =>
 interface PlaylistDetailProps {
   playlistId: string;
   userId?: string;
-  playlistData: GetPlaylistQuery;
+  playlistData: PlaylistJson[];
+  playlistName: string;
 }
 
 export default function PlaylistDetail({
   playlistId,
   userId,
   playlistData,
+  playlistName,
 }: PlaylistDetailProps) {
   const { toast } = useToast();
   const { token, user } = useAuthStore();
 
-  const sanitizedPlaylistJson = useSanitizedData(
-    playlistData?.playlist.listJson
-  );
   if (!playlistId) return <div>Playlist not found</div>;
 
   const handleDownloadCSV = () => {
     try {
-      if (!sanitizedPlaylistJson) return;
+      if (!playlistData) return;
       const headerString = HEADERS.map((header) => `"${header}"`).join(",");
 
-      const rows = sanitizedPlaylistJson.reduce((acc, song) => {
+      const rows = playlistData.reduce((acc, song) => {
         return (
           acc +
           "\n" +
@@ -53,9 +50,7 @@ export default function PlaylistDetail({
         );
       }, headerString);
 
-      const sanitizedFileName = sanitizeFileName(
-        playlistData?.playlist.name || ""
-      );
+      const sanitizedFileName = sanitizeFileName(playlistName);
 
       const blob = new Blob([BOM + rows], {
         type: "text/csv;charset=utf-8",
@@ -89,17 +84,17 @@ export default function PlaylistDetail({
           isBelongsToUser={userId === user?.id}
           token={token}
           playlistId={playlistId}
-          playlistName={playlistData?.playlist.name || ""}
+          playlistName={playlistName}
         />
       </div>
 
       <div className="flex flex-col md:flex gap-4 w-full">
         <ConvertToYoutubePlaylistButton
-          playlistData={sanitizedPlaylistJson || []}
+          playlistData={playlistData || []}
           token={token}
         />
         <ConvertToSpotifyPlaylistButton
-          playlistData={sanitizedPlaylistJson || []}
+          playlistData={playlistData || []}
           token={token}
         />
         <Button
@@ -112,8 +107,6 @@ export default function PlaylistDetail({
           Download CSV
         </Button>
       </div>
-
-      <SongTable songs={sanitizedPlaylistJson || []} />
     </>
   );
 }
