+++ src/app/playlists/[playlist-id]/page.tsx
@@ -3,6 +3,9 @@ import PlaylistDetail from "./_components/playlist";
 import { GetPlaylistDocument } from "@/graphql/hooks";
 import { GetPlaylistQuery } from "@/graphql/operations";
 import NotFound from "@/app/not-found";
+import ApolloProvider from "@/providers/apollo-provider";
+import { sanitizeData } from "@/lib/sanitize-data";
+import { SongTable } from "../_components/song-table";
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
 export default async function PlaylistDetailPage(props: any) {
@@ -12,20 +15,23 @@ export default async function PlaylistDetailPage(props: any) {
     query: GetPlaylistDocument,
     variables: { id: parseInt(playlistId) },
   });
+  const sanitizedData = sanitizeData(data?.playlist.listJson);
 
   if (error) {
     return <NotFound />;
   }
 
   return (
-    <div className="container mx-auto px-4 py-8">
-      <div className="space-y-6 w-full">
+    <>
+      <ApolloProvider>
         <PlaylistDetail
           playlistId={playlistId}
           userId={data?.playlist.userId}
-          playlistData={data}
+          playlistData={sanitizedData || []}
+          playlistName={data?.playlist.name}
         />
-      </div>
-    </div>
+      </ApolloProvider>
+      <SongTable songs={sanitizedData || []} />
+    </>
   );
 }
