+++ src/app/playlists/layout.tsx
@@ -1,4 +1,5 @@
 import Navbar from "@/components/nav-bar";
+import ApolloProvider from "@/providers/apollo-provider";
 
 interface PlaylistsLayoutProps {
   children: React.ReactNode;
@@ -9,7 +10,9 @@ export default async function PlaylistsLayout({
 }: PlaylistsLayoutProps) {
   return (
     <div className="min-h-screen">
-      <Navbar />
+      <ApolloProvider>
+        <Navbar />
+      </ApolloProvider>
       {children}
     </div>
   );
