+++ src/app/statistics/layout.tsx
@@ -1,5 +1,5 @@
 import Navbar from "@/components/nav-bar";
-
+import ApolloProvider from "@/providers/apollo-provider";
 export default function StatisticsLayout({
   children,
 }: {
@@ -7,8 +7,10 @@ export default function StatisticsLayout({
 }) {
   return (
     <div className="min-h-screen">
-      <Navbar />
-      {children}
+      <ApolloProvider>
+        <Navbar />
+        {children}
+      </ApolloProvider>
     </div>
   );
 }
